spring-mongodb-testcontainers/src/main/java/com/example/springmongodbtestcontainers/service/ProductService.java
package com.example.springmongodbtestcontainers.service;

import com.example.springmongodbtestcontainers.entity.Product;
import com.example.springmongodbtestcontainers.repository.ProductRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductService {
    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public Product save(Product product) {
        return productRepository.save(product);
    }

    public Product findById(String id) {
        return productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found"));
    }

    public List<Product> findAll() {
        return productRepository.findAll();
    }

    public void deleteById(String id) {
        productRepository.deleteById(id);
    }
}
